Create db.ts
Added the line relationMode = "prisma" in the prisma.schema file. Then, created a file named db.ts in the lib folder.

In the db.ts file:

Declared a global variable named prisma with the type PrismaClient | undefined, allowing access to the Prisma client instance from anywhere in the codebase.
Initialized the prisma variable with the initial value of globalThis.prisma, which is expected to be undefined on the first run. If prisma is undefined, a new PrismaClient instance is created and assigned to prisma.
Utilized an if statement to check if the environment is not production. If so, assigned the db constant back to the globalThis.prisma variable, effectively making it a globally accessible singleton instance.