Update queries.ts
Created the createTeamUser query. Here's an overview of its purpose and functionality:

Purpose: The query is responsible for creating a new user within the agency.

Functionality:

First, it checks if the currently logged-in user is an agency owner.
If the logged-in user is an agency owner, the db.user.create method is used to create a new user.