Update queries.ts
The upsertAgency query is designed to add or update agency information in the database. It facilitates the creation of new agencies or the modification of existing ones, ensuring seamless management of agency data within the system.

Implementation Details:

Defining upsertAgency Query:
The upsertAgency query is implemented in the lib/queries.ts file.
It accepts two parameters: agency, representing the agency information to be inserted or updated, and price (optional), denoting the subscription plan associated with the agency.
The query first checks if the agency's company email exists. If not, it returns null, indicating invalid agency data.
Within the main try-catch block, the query proceeds with database operations.
Using the upsert method provided by Prisma, the query attempts to update the agency information based on the agency ID.
If the agency ID does not exist, the query creates a new agency entry in the database.
The upsert method is configured with the following parameters:
where: Specifies the condition for finding the agency based on its ID.
update: Defines the data to be updated for an existing agency.
create: Specifies the data to be inserted if the agency does not exist.
When creating a new agency, the query also associates users with the agency by connecting them based on their email addresses.
Additionally, the query sets up SidebarOptions for the agency, providing predefined navigation links for various sections such as Dashboard, Launchpad, Billing, Settings, Sub Accounts, and Team.
The function returns the agencyDetails, representing the updated agency information after completion.